Fix wrong redirect following authorization step, refs #171
Fix for #163 (https://github.com/bshaffer/oauth2-server-php/issues/163) allowed
authorize flow to handle `redirect_uri` as an optional parameter. The change
resulted in user not being redirected back to the client after granting or
denying client's access.

This fix verifies whether `redirect_uri` is set or not in
`OAuth2\Controller\AuthorizeController`'s `handleAuthorizeRequest()` method.
If it's not set, it grabs the client's registered value.

If/When support for PHP 5.3 is dropped, the code can be modified to
`$this->clientStorage->getClientDetails($this->client_id)['redirect_uri']`.